Bump minimum code coverage threshold for most projects
The "app" project is an outlier at this time but will hopefully
have its coverage improved in the future.  However, it shouldn't cause
the other subprojects to have an artificially low minimum threshold.

The manner in which a different threshold is implemented for the "app"
project smells a bit.  The problem is configuring the
JacocoCoverageVerification task in the "app" project simply adds a
second rule without overriding the first.  Going through the trouble to
find the original rule defined in the root project and overwriting its
limit is possible, but it seems to run after the build script has been
configured.  Thus, the change is not actually processed by Gradle, and
the original threshold is used.  Setting a different threshold based on
the project name in the root build script is ugly but it works as
expected.