Remove GridTypeRegistry and UnitTypeRegistry
These types are not applicable for the game-core project.  They should
be in a separate "services" module, or at least the "app" module.  As
they are used simply by the Training Scenario tests, they will be
removed for now, and the required GridType and MechType instances
inlined into the test fixture.