Require explicit trigger to end deployment phase
In fact, all phases now require an explicit trigger to end them and move
on to the next phase.  This was done to avoid prematurely moving to the
next phase in the event a player wants to perform another action in the
active phase.  For example, after the last Mech is deployed, the player
may choose to change the deployment position before ending the phase
rather than having the game automatically ending the phase once all
Mechs are deployed.