Extract phase implementations to standalone classes
To achieve this, the game state is now stored in a separate type
(DefaultGameState) instead of DefaultGame directly.

This also allows the phase specs to be written independently of the
phase state machine.  That is, a spec doesn't have to assume it always
starts with the initialization phase and must progress to the phase
under test, which would only get uglier as more phases are added.
Instead, the spec now initializes the game to a desired state before
invoking methods of the phase under test.