Result$Failure should not be a data class
Throwable is not equatable/hashable.  Thus, it doesn't make sense for
Result$Failure to be equatable/hashable.

We could try to use some heuristic to determine if two exceptions are
equivalent (e.g. type and message), but this feature is probably a YAGNI
at this time, and I'm not even sure if it makes sense.