Upgrade ktlint
The latest version of ktlint adds a rule to enforce the Kotlin style
guide filename conventions.  We have several violations in our test code
due to our attempt to keep the spec files named after the corresponding
production file.

In a few cases, the specs were restructured in a more sensible way that
automatically fixed the filename rule violation.  In the majority of
cases, a workaround was used whereby a no-op spec (with the same name as
the file) was added to satisfy the filename rule.  These no-op specs
fall into two categories:

* The only other spec in the file is a DataClassSpec.  DataClassSpec
  will be removed in the future now that JaCoCo correctly ignores
  generated Kotlin code.  (This is currently an unreleased feature of
  JaCoCo.)

* Second and subsequent specs are anticipated to be added to the file
  (e.g. command specs), at which time the placeholder spec can be
  removed.