Temporarily disable ktlint filename rule
This partially reverts the placeholder specs added in c5d494e.  Instead
of using the placeholder specs, we now simply disable the associated
rule.  The same circumstances still apply as in the previous commit,
namely that disabling this rule is anticipated to only be temporary, and
these directives will be removed in the future.