Update IDEA module configuration
This probably should have been part of either the upgrade to
IDEA 2018.3 or Kotlin 1.3, but, for some reason, IDEA didn't flush the
module configuration to disk.

I'm not sure why the classpath contains user-specific folders, where
that's never been an issue in previous versions of IDEA.  I'm hoping
that will be addressed in a future IDEA update.  Otherwise, I'll
manually replace the hard-coded paths with the appropriate macro.